email_communicator-agent/pyproject.toml
[project]
name = "email-communicator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "fastapi[standard]>=0.115.14",
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "grpcio<1.72.1",
    "openai==1.96.1",
    "openai-agents>=0.2.3",
    "pytz>=2025.2",
    "uvicorn>=0.35.0",
]
